server/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlmodel==0.0.19
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── Graph DB ─────────────────────────────────────────────────────────────────
neo4j==5.19.0

# ─── Vector DB ────────────────────────────────────────────────────────────────
qdrant-client==1.9.1

# ─── Job queue ────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.4
flower==2.0.1

# ─── Parsing ──────────────────────────────────────────────────────────────────
tree-sitter==0.23.0
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
tree-sitter-go==0.23.0
tree-sitter-rust==0.23.0
tree-sitter-java==0.23.0
gitpython==3.1.43
pathspec==0.12.1

# ─── Embedding / LLM ──────────────────────────────────────────────────────────
google-generativeai==0.7.2
voyageai==0.3.2
tiktoken==0.7.0

# ─── Graph algorithms ─────────────────────────────────────────────────────────
networkx==3.3
python-louvain==0.16

# ─── Utilities ────────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
aiofiles==23.2.1
tenacity==8.3.0
structlog==24.1.0
rich==13.7.1
